converted server js to ts
