+++ package.json
@@ -18,6 +18,8 @@
     "@babel/preset-react": "^7.23.3",
     "@types/archiver": "^6.0.2",
     "@types/aws-sdk": "^2.7.0",
+    "@types/cookie-parser": "^1.4.6",
+    "@types/cors": "^2.8.17",
     "@types/dotenv": "^8.2.0",
     "@types/express": "^4.17.21",
     "@types/node": "^20.11.10",
