+++ tsconfig.json
@@ -9,11 +9,7 @@
     "moduleResolution": "node",
     "outDir": "./build"
   },
-  "files": [
-    "docs/types.ts",
-    "server/server.ts",
-    "server/controllers/fileController.ts"
-  ],
+  "files": ["docs/types.ts"],
   "exclude": ["**/*.js"],
-  "include": ["client/utils/**/*"]
+  "include": ["client/utils/**/*", "server/**/*"]
 }
